Split cloud backup manager modules
